+++ CHANGELOG.md
@@ -1,5 +1,7 @@
+## 0.0.5
+- add border border-radius bg-gradient box-shadow
 ## 0.0.4
-- add textAlign , grid ,aspectRadio
+- add textAlign grid grid-cols aspectRadio
 ## 0.0.3
 - add lineHeight maxLine fontStyle
 ## 0.0.2